+++ HISTORY.rst
@@ -10,10 +10,18 @@ History
 * Drop support for Django < 1.8
 * Drop support for django CMS < 3.2
 
-0.8.8 (unreleased)
+0.8.9 (2016-05-19)
+++++++++++++++++++
+
+* Add support for django CMS 3.3rc
+
+0.8.8 (2016-05-06)
 ++++++++++++++++++
 
 * Force language codes to lowercase
+* Force i18n if multiple languages is provided
+* Fix some errors in selecting dependencies
+* Fix error in Django 1.9 regexp
 
 0.8.7 (2016-02-23)
 ++++++++++++++++++